+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Store the browser authentication secret on Web Push subscription instead of the channel authentication token, fixing the message decryption on the client
 
 ## [0.6.8] - 2026-06-28